feat: add functions for displaying and hiding flex and justify drop-down lists
