+++ src/components/showJustifyContentDropdown.ts
@@ -1,0 +1,7 @@
+export function showJustifyContentDropdown(): void {
+    const JustifyContentDropdown = document.getElementById("container-justify-content");
+    if (JustifyContentDropdown) {
+      JustifyContentDropdown.style.display = "flex";
+      console.info("container-justify-content для show нашёлся") ;}
+    else {console.info("container-justify-content для show not found");}
+  }
